product services completed

+++ customerapp/models/customers.model.ts
@@ -9,28 +9,27 @@ class customerModelRegistery {
   public models: { [key: string]: ModelStatic<Model<any, any>> } = {};
   private constructor() {}
 
-  
   public static getInstance(): customerModelRegistery {
     if (!customerModelRegistery.instance) {
-      return (customerModelRegistery.instance = new customerModelRegistery());
+      return customerModelRegistery.instance = new customerModelRegistery();
     }
       return customerModelRegistery.instance;
     
   }
 
   public async initModel(): Promise<void> {
-    try {
-      const response = await apiRequest<ModelAttributes<Model>>(
-        "GET",
-        "http://localhost:3009/customer-model",
-        undefined,
-        undefined,
-        {}
-      );
-      this.models.customers = sequelize.define("customers", response.data);
-      await sequelize.sync({force:false,alter: true });
-    } catch (error) {
-      console.log("customerModel api get error--->",error);
+    if (!this.models.customers) {
+      try {
+        const response = await apiRequest<ModelAttributes<Model>>(
+          "GET",
+          "http://localhost:3009/customer-model"
+        );
+    
+        this.models.customers = sequelize.define("customers", response.data);
+        await sequelize.sync({ force: false, alter: true });
+      } catch (error) {
+        console.log("customerModel api get error--->", error);
+      }
     }
   }
